+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 3.0.6 (unreleased)
 ------------------
 
+- fix failing unpickling of older Z SQ Methods instances
+  (`#12 <https://github.com/zopefoundation/Products.ZSQLMethods/issues/12>`_)
+
 - improve usability and add Zope 4 compatibility for connection forms
 
 - improve general usability of the ZSQL Method edit and test forms